+++ src/Localization/_LOCALIZATION_MODULE.mjs
@@ -1,1 +1,0 @@
-export const LOCALIZATION_LOCALIZATION_MODULE = "localization";