+++ src/SelectLanguage/setLanguage.mjs
@@ -1,3 +1,0 @@
-/**
- * @typedef {(language: string | null) => void} setLanguage
- */